fix(viewer): resolve annotation hit-test padding vs paint-order conflict; shrink useAnnotation2D under budget
CI failures on #4198:

- check-module-size.mjs: useAnnotation2D.ts grew to 562 lines against a
  551-line budget. Extracted hit-testing into a new, independently
  testable apps/viewer/src/hooks/annotation2DHitTest.ts, bringing the
  hook to 461 lines; lowered its allowlist budget to match.

- check-unused-locals.mjs: the new hitTestOrder test file imported
  `before` from node:test without using it, tripping the ratchet for
  apps/viewer and apps/viewer-embed (the latter typechecks the former's
  sources via its tsconfig paths). Removed the unused import.

Also fixes a real regression this PR introduced while it was here: a
cloud's hit-test bbox is padded by HIT_TEST_RADIUS_PX for click
tolerance, so under the reverse-paint-order fix (cloud checked first,
since it paints last) a click that only lands in that padding — not the
cloud's actual drawn extent — could steal a click meant for a genuinely
underlying annotation. Hit-testing now resolves in two passes: genuine
hits (inside a shape's true rendered extent) win by paint order, and
only when no genuine hit exists does nearest-distance among padded
(tolerance-only) hits decide. Added a regression test for the exact
padding-vs-real-hit fixture and mutation-tested it (reverting the pass-1
padding removal turns the new test red; the other 5 stay green).
